REFACTOR(RestApi proj): moved configurators to separate namespace for better comprehension
