+++ OnlineMinion.RestApi/Configuration/RouteOptionsConfigurator.cs
@@ -2,7 +2,7 @@ using JetBrains.Annotations;
 using Microsoft.AspNetCore.Routing;
 using Microsoft.Extensions.Options;
 
-namespace OnlineMinion.RestApi.Init;
+namespace OnlineMinion.RestApi.Configuration;
 
 [UsedImplicitly]
 public class RouteOptionsConfigurator : IConfigureOptions<RouteOptions>
